separate logic for Auth and User
