+++ CineApi/Controllers/AuthController.cs
@@ -91,89 +91,5 @@ namespace CineApi.Controllers
                 return StatusCode(500, new { message = AuthValidationMessages.InternalServerError() });
             }
         }
-
-        [HttpGet("users")]
-        [Authorize(Roles = UserRoles.SysAdmin)]
-        public async Task<IActionResult> GetAllUsers()
-        {
-            try
-            {
-                var users = await _authService.GetAllUsers();
-                return Ok(users);
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, new { message = AuthValidationMessages.InternalServerError() });
-            }
-        }
-
-        [HttpPost("users")]
-        [Authorize(Roles = UserRoles.SysAdmin)]
-        public async Task<IActionResult> CreateUser([FromBody] CreateUserDto createUserDto)
-        {
-            try
-            {
-                if (!ModelState.IsValid)
-                {
-                    return BadRequest(ModelState);
-                }
-
-                var result = await _authService.CreateUser(createUserDto);
-                return Ok(result);
-            }
-            catch (InvalidOperationException ex)
-            {
-                return BadRequest(new { message = ex.Message });
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, new { message = AuthValidationMessages.InternalServerError() });
-            }
-        }
-
-        [HttpPut("users/{id}")]
-        [Authorize(Roles = UserRoles.SysAdmin)]
-        public async Task<IActionResult> UpdateUser(int id, [FromBody] UpdateUserDto updateUserDto)
-        {
-            try
-            {
-                if (!ModelState.IsValid)
-                {
-                    return BadRequest(ModelState);
-                }
-
-                var result = await _authService.UpdateUser(id, updateUserDto);
-                if (result == null)
-                {
-                    return NotFound(new { message = AuthValidationMessages.UserNotFound() });
-                }
-
-                return Ok(result);
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, new { message = AuthValidationMessages.InternalServerError() });
-            }
-        }
-
-        [HttpDelete("users/{id}")]
-        [Authorize(Roles = UserRoles.SysAdmin)]
-        public async Task<IActionResult> DeleteUser(int id)
-        {
-            try
-            {
-                var success = await _authService.DeleteUser(id);
-                if (!success)
-                {
-                    return NotFound(new { message = AuthValidationMessages.UserNotFound() });
-                }
-
-                return Ok(new { message = AuthValidationMessages.UserDeletedSuccessfully() });
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, new { message = AuthValidationMessages.InternalServerError() });
-            }
-        }
     }
 }
